python/requirements.txt
pyserial
pandas
matplotlib
fpdf2